requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
croniter==6.0.0
distro==1.9.0
dnspython==2.8.0
fastapi==0.128.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.81
langchain-openai==0.2.14
langgraph==0.2.58
langgraph-checkpoint==2.1.2
langgraph-sdk==0.1.74
langsmith==0.6.1
openai==1.58.1
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pdf2image==1.17.0
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.15.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rq==2.6.1
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.40.0
zstandard==0.25.0
